+++ test/backend/dyn_replace_slice_reference.in.cpp
